+++ fe/src/components/common/SheetSide.tsx
@@ -1,58 +1,61 @@
 'use client';
 
 import {
-  Drawer,
-  DrawerClose,
-  DrawerContent,
-  DrawerDescription,
-  DrawerFooter,
-  DrawerHeader,
-  DrawerTitle,
-  DrawerTrigger,
-} from '@/components/ui/drawer';
+  Sheet,
+  SheetClose,
+  SheetContent,
+  SheetDescription,
+  SheetFooter,
+  SheetHeader,
+  SheetTitle,
+  SheetTrigger,
+} from '@/components/ui/sheet';
 
-import { Button } from '@/components/ui/button';
+import { Button } from '../ui/button';
+import { ChevronsRight } from 'lucide-react';
 import Link from 'next/link';
 import { ScrollArea } from '@/components/ui/scroll-area';
 import { Transaction } from '@/types/testTransaction';
 import TransactionSection from '@/components/common/TransactionSection';
-import { X } from 'lucide-react';
 
-export function DrawerBottom({ data }: { data: Transaction[] }) {
+const SheetSide = ({ data }: { data: Transaction[] }) => {
   const totalCount = data.length;
 
   return (
-    <Drawer>
-      <DrawerTrigger>
-        <span className="cursor-pointer">Open Drawer Bottom</span>
-      </DrawerTrigger>
+    <Sheet>
+      <SheetTrigger>
+        <span className="cursor-pointer">Open Sheet Side</span>
+      </SheetTrigger>
 
-      <DrawerContent className="h-[85vh] outline-none">
+      <SheetContent
+        side="right"
+        className="flex w-[540px] max-w-none flex-col outline-none sm:max-w-none"
+      >
         {/* 헤더 영역: 날짜 및 닫기 버튼 */}
-        <DrawerHeader className="relative flex shrink-0 flex-row items-center justify-between border-b px-6 py-4">
-          <DrawerClose className="absolute top-0 right-0 mr-4" asChild>
+        <SheetHeader className="flex shrink-0 justify-between border-b p-4">
+          <SheetClose asChild>
             <Button
               variant="ghost"
               size="icon"
               className="h-10 w-10 cursor-pointer rounded-full"
             >
-              <X />
+              <ChevronsRight size={20} />
             </Button>
-          </DrawerClose>
+          </SheetClose>
 
-          <div className="text-left">
-            <DrawerTitle className="py-2 text-xl font-bold">
+          <div className="pl-2 text-left">
+            <SheetTitle className="py-2 text-xl font-bold">
               12월 24일 수요일
-            </DrawerTitle>
-            <DrawerDescription>
+            </SheetTitle>
+            <SheetDescription>
               해당 날짜 거래 내역을 확인할 수 있습니다.
-            </DrawerDescription>
+            </SheetDescription>
           </div>
 
-          <div className="flex h-full items-end">
-            <DrawerDescription>총 {totalCount}개</DrawerDescription>
-          </div>
-        </DrawerHeader>
+          <SheetDescription className="text-right">
+            총 {totalCount}개
+          </SheetDescription>
+        </SheetHeader>
 
         {/* 콘텐츠 영역: 거래 내역 영역 */}
         <ScrollArea className="flex-1 overflow-y-auto">
@@ -66,7 +69,7 @@ export function DrawerBottom({ data }: { data: Transaction[] }) {
         </ScrollArea>
 
         {/* 푸터 영역: 이번 달 거래 내역 버튼 */}
-        <DrawerFooter className="shrink-0 border-t p-6">
+        <SheetFooter className="shrink-0 border-t p-6">
           <Button
             variant="outline"
             className="h-14 w-full cursor-pointer text-base"
@@ -74,8 +77,10 @@ export function DrawerBottom({ data }: { data: Transaction[] }) {
           >
             <Link href="/history">이번 달 거래 내역 보러가기</Link>
           </Button>
-        </DrawerFooter>
-      </DrawerContent>
-    </Drawer>
+        </SheetFooter>
+      </SheetContent>
+    </Sheet>
   );
-}
+};
+
+export default SheetSide;
